+++ tests/e2e/create/create_suite_test.go
@@ -4,8 +4,6 @@ package create_test
 
 import (
 	"fmt"
-	"io/fs"
-	"os"
 	"os/exec"
 	"strings"
 	"testing"
@@ -85,42 +83,4 @@ func (cmd *createCmd) execute() error {
 		return fmt.Errorf("create command failed with output: %s and error: %w", cmdOut, err)
 	}
 	return nil
-}
-
-func filesIn(dir string) []string {
-	fi, err := os.Stat(dir)
-	Expect(err).ToNot(HaveOccurred())
-	Expect(fi.IsDir()).To(BeTrueBecause("The provided path should be a directory: %s", dir))
-
-	dirFs := os.DirFS(dir)
-	entries, err := fs.ReadDir(dirFs, ".")
-	Expect(err).ToNot(HaveOccurred())
-
-	var res []string
-	for _, ent := range entries {
-		if ent.Type().IsRegular() {
-			res = append(res, ent.Name())
-		}
-	}
-
-	return res
-}
-
-func resolveWorkingDirectory() string {
-	createDir := os.Getenv("CREATE_DIR")
-	if len(createDir) > 0 {
-		return createDir
-	}
-
-	createDir, err := os.MkdirTemp("", "create_test")
-	if err != nil {
-		Fail(err.Error())
-	}
-	return createDir
-}
-
-func cleanupWorkingDirectory(path string) {
-	if len(os.Getenv("CREATE_DIR")) == 0 {
-		_ = os.RemoveAll(path)
-	}
 }
